requirements.txt
python-telegram-bot==22.4
APScheduler==3.10.4
requests==2.32.3
openai>=1.30.0
pytz==2024.1
psutil==5.9.8
feedparser==6.0.11
